requirements.txt
flask
tensorflow
numpy
opencv-python
pygame
python-nmap
tf_keras
